Tests compile, but 7 fail

+++ src/tests.rs
@@ -25,7 +25,7 @@ use sp_runtime::MultiSignature;
 
 // Constant that reflects the desired vesting period for the tests
 // Most tests complete initialization passing initRelayBlock + VESTING as the endRelayBlock
-const VESTING: u32 = 8;
+const VESTING: u64 = 8;
 
 #[test]
 fn geneses() {
